Align auth imports with main
