+++ client/src/components/CreateTournament.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { QRCodeSVG } from "qrcode.react";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../contexts/useAuth";
 import { createTournament } from "../api/tournaments";
 import type { Tournament } from "../api/tournaments";
