+++ client/src/pages/LoginPage.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../contexts/useAuth";
 import { checkBootstrap } from "../api/auth";
 
 export default function LoginPage() {
